// projects/anyopsos-app-datastore-explorer/src/lib/types/datastore-explorer-connection.ts
import {ImDataObject, NetAppVolume, VMWareDatastore, VMWareDatacenter} from '@anyopsos/app-infrastructure-manager';

export interface DatastoreExplorerConnection {
  uuid?: string;
  credential: string;
  host: string;
  port: number;
  type: string;
  data: {
    obj: ImDataObject & { info: { data: VMWareDatastore | NetAppVolume } };
    datacenter?: ImDataObject & { info: { data: VMWareDatacenter } };
  };
  state?: string;
}